wrap up the new filter view
